Simplify GitHub Action workflow
The commit drastically simplifies the GitHub Actions workflow, removing numerous steps previously involved in the process such as checking out the repository, setting up Python, installing requirements, etc. This leaves only the configuration of AWS credentials as the remaining task. The AWS region is now hard-coded to "sa-east-1" instead of being a variable.